requirements.txt
Flask
gunicorn
mysql-connector-python
sentence-transformers
faiss-cpu
numpy